refactor(shopping card): move ShoppingCardProps into separate file
