+++ src/Components/ShoppingCard/ShoppingCard.d.ts
@@ -1,0 +1,15 @@
+type ShoppingCardProps = {
+  className?: string
+  name: string
+  description: string
+  imageUrl: string
+  imageAlt: string
+  price: number
+  currency: string
+  discountRate?: number
+  onNameClick?: () => void
+  toFixed?: 2
+  intlLocale?: string
+}
+
+export default ShoppingCardProps
